# OpenAPI.yaml
openapi: 3.0.1
info:
  title: API Vulpes
  description: ""
  contact:
    name: Kaique de Miranda
    email: kbdemiranda@hotmail.com
servers:
- url: http://127.0.0.1:8080/api
  description: Generated server url
...
paths:
...
  /plataformas/{id}:
    get:
      tags:
      - Plataformas
      summary: Buscar plataforma
      description: Retorna uma plataforma pelo ID
      operationId: buscarPlataforma
      parameters:
      - name: id
        in: path
        description: ID da plataforma
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Plataforma encontrada
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plataforma'
        "404":
          description: Plataforma não encontrada
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer-key: []
    put:
      tags:
      - Plataformas
      summary: Atualizar plataforma
      description: Atualiza dados de uma plataforma existente
      operationId: atualizarPlataforma
      parameters:
      - name: id
        in: path
        description: ID da plataforma
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Plataforma'
        required: true
      responses:
        "404":
          description: Plataforma não encontrada
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "201":
          description: Plataforma atualizada
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plataforma'
      security:
      - bearer-key: []
    delete:
      tags:
      - Plataformas
      summary: Excluir plataforma
      description: Remove uma plataforma pelo ID
      operationId: excluirPlataforma
      parameters:
      - name: id
        in: path
        description: ID da plataforma
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "404":
          description: Plataforma não encontrada
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "204":
          description: Plataforma excluída
      security:
      - bearer-key: []
  /pagamentos/{id}:
    get:
      tags:
      - Pagamentos
      summary: Consultar pagamento
      description: Recupera um pagamento pelo ID
      operationId: consultarPagamento
      parameters:
      - name: id
        in: path
        description: ID do pagamento
        required: true
        schema:
          type: integer
          format: int64
        example: 10
      responses:
        "200":
          description: Pagamento encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Pagamento'
        "404":
          description: Pagamento não encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer-key: []
    put:
      tags:
      - Pagamentos
      summary: Atualizar pagamento
      description: Atualiza um pagamento existente
      operationId: atualizarPagamento
      parameters:
      - name: id
        in: path
        description: ID do pagamento
        required: true
        schema:
          type: integer
          format: int64
        example: 10
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pagamento'
        required: true
      responses:
        "201":
          description: Pagamento atualizado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Pagamento'
        "404":
          description: Pagamento não encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer-key: []
...
  /plataformas:
    get:
      tags:
      - Plataformas
      summary: Listar plataformas
      description: "Retorna lista paginada de plataformas, com filtro opcional por\
        \ nome"
      operationId: listarPlataformas
      parameters:
      - name: nome
        in: query
        description: Filtro por nome
        required: false
        schema:
          type: string
        example: Netflix
      - name: pagina
        in: query
        description: Número da página (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
        example: 0
      - name: quantidade
        in: query
        description: Quantidade por página
        required: false
        schema:
          minimum: 1
          type: integer
          format: int32
          default: 10
        example: 10
      responses:
        "500":
          description: Erro interno
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Lista retornada com sucesso
          content:
            application/json: {}
        "401":
          description: Não autorizado
      security:
      - bearer-key: []
    post:
      tags:
      - Plataformas
      summary: Cadastrar plataforma
      description: Cria uma nova plataforma
      operationId: cadastrarPlataforma
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Plataforma'
        required: true
      responses:
        "201":
          description: Plataforma criada
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plataforma'
        "400":
          description: Dados inválidos
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer-key: []
  /pagamentos:
    get:
      tags:
      - Pagamentos
      summary: Listar pagamentos
      description: "Retorna lista paginada de pagamentos, com filtro opcional por\
        \ nome do assinante"
      operationId: listarPagamentos
      parameters:
      - name: nomeAssinante
        in: query
        description: Filtro por nome do assinante
        required: false
        schema:
          type: string
        example: João
      - name: pagina
        in: query
        description: Número da página (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
        example: 0
      - name: quantidade
        in: query
        description: Quantidade por página
        required: false
        schema:
          minimum: 1
          type: integer
          format: int32
          default: 10
        example: 10
      responses:
        "500":
          description: Erro interno
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Lista retornada com sucesso
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Não autorizado
      security:
      - bearer-key: []
    post:
      tags:
      - Pagamentos
      summary: Registrar pagamento
      description: Cria um novo pagamento
      operationId: registrarPagamento
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pagamento'
        required: true
      responses:
        "201":
          description: Pagamento criado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Pagamento'
        "400":
          description: Dados inválidos
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer-key: []
...
  /pagamentos/assinante/{id}:
    get:
      tags:
      - Pagamentos
      summary: Listar pagamentos por assinante
      description: Retorna lista paginada de pagamentos de um assinante
      operationId: listarPagamentosAssinante
      parameters:
      - name: id
        in: path
        description: ID do assinante
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      - name: pagina
        in: query
        description: Número da página (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
        example: 0
      - name: quantidade
        in: query
        description: Quantidade por página
        required: false
        schema:
          minimum: 1
          type: integer
          format: int32
          default: 10
        example: 10
      responses:
        "404":
          description: Assinante não encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Lista retornada com sucesso
          content:
            '*/*':
              schema:
                type: object
      security:
      - bearer-key: []
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          description: Data/hora do erro (UTC)
          format: date-time
        status:
          type: integer
          description: Código HTTP do erro
          format: int32
          example: 404
        error:
          type: string
          description: Tipo de erro
          example: VulpesException
        message:
          type: string
          description: Mensagem explicativa do erro
          example: Plataforma não encontrada
        path:
          type: string
          description: Caminho da requisição
          example: /plataformas/999
      description: Estrutura padronizada de erros da API
    Plataforma:
      required:
      - nome
      - preco
      - tipo_servico
      - total_vagas
      type: object
      properties:
        id:
          type: integer
          description: Identificador da plataforma
          format: int64
          example: 1
        nome:
          maxLength: 255
          minLength: 3
          type: string
          description: Nome da plataforma
          example: Netflix
        preco:
          minimum: 0.0
          exclusiveMinimum: true
          type: number
          description: Preço mensal
          example: 29.9
        url:
          type: string
          description: URL oficial
          example: https://netflix.com
        tipo_servico:
          type: string
          description: Tipo de serviço de uma plataforma
          example: STREAMING_VIDEO
          enum:
          - STREAMING_VIDEO
          - STREAMING_MUSICA
          - SOFTWARE
          - JOGOS
          - NOTICIAS
          - CLOUD_STORAGE
          - FITNESS
        total_vagas:
          minimum: 1
          type: integer
          description: Total de vagas
          format: int32
          example: 5
        vagas_disponiveis:
          type: integer
          description: Vagas disponíveis
          format: int32
          example: 5
        cadastrado_em:
          type: string
          description: Data de criação
          format: date-time
        atualizado_em:
          type: string
          description: Data de atualização
          format: date-time
        excluido_em:
          type: string
          description: Data de exclusão (soft delete)
          format: date-time
      description: Dados de uma plataforma de serviço
    Pagamento:
      required:
      - assinante_id
      - data_pagamento
      - meses_cobertos
      - valor_pago
      type: object
      properties:
        id:
          type: integer
          description: Identificador do pagamento
          format: int64
          example: 10
        assinante_id:
          type: integer
          description: ID do assinante
          format: int64
          example: 1
        valor_pago:
          minimum: 0.01
          exclusiveMinimum: false
          type: number
          description: Valor pago
          example: 59.8
        meses_cobertos:
          minimum: 1
          type: integer
          description: Quantidade de meses cobertos
          format: int32
          example: 2
        meses:
          type: array
          description: Meses de referência
          example:
          - 2024-07
          - 2024-08
          items:
            type: string
            description: Meses de referência
            example: "[\"2024-07\",\"2024-08\"]"
        data_pagamento:
          type: string
          description: Data do pagamento
          format: date-time
        cadastrado_em:
          type: string
          description: Data de criação
          format: date-time
        atualizado_em:
          type: string
          description: Data de atualização
          format: date-time
        excluido_em:
          type: string
          description: Data de exclusão (soft delete)
          format: date-time
      description: Dados de um pagamento
...
  securitySchemes:
    bearer-key:
      type: http
      scheme: bearer
      bearerFormat: JWT